+++ buttons/verifyme.js
@@ -4,6 +4,8 @@
  * See LICENSE for license details
  */
 
+const { applyCitizenship } = require('../helpers/customs.js');
+
 module.exports = {
     data: {name: 'verifyme'},
     hasPlayer: true,